fix warning.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@22297 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/insets/InsetSpace.cpp
@@ -69,7 +69,7 @@ void InsetSpace::metrics(MetricsInfo & mi, Dimension & dim) const
 			break;
 		case ENSPACE:
 		case ENSKIP:
-			dim.wid = 0.5 * fm.width(char_type('M'));
+			dim.wid = int(0.5 * fm.width(char_type('M')));
 			break;
 	}
 	// Cache the inset dimension.
